requirements.txt
streamlit
requests
gtts==2.2.4
pillow
